# GH-4012 use toStringCache in MemIRI .equals(...) 