@hmottestad 2022-06-23
## Before
```
Benchmark                                    Mode  Cnt   Score   Error  Units
ParallelQueryBenchmark.mixedQueriesAndReads  avgt   15  29.058 ± 0.290  ms/op
```

```
Benchmark                                                     Mode  Cnt     Score    Error  Units
QueryBenchmark.complexQuery                                   avgt    5     1.755 ±  0.014  ms/op
QueryBenchmark.different_datasets_with_similar_distributions  avgt    5     0.844 ±  0.003  ms/op
QueryBenchmark.groupByQuery                                   avgt    5     1.085 ±  0.007  ms/op
QueryBenchmark.long_chain                                     avgt    5   306.587 ±  1.005  ms/op
QueryBenchmark.lots_of_optional                               avgt    5    61.701 ±  0.168  ms/op
QueryBenchmark.minus                                          avgt    5  2396.788 ± 46.319  ms/op
QueryBenchmark.nested_optionals                               avgt    5    85.223 ±  0.195  ms/op
QueryBenchmark.pathExpressionQuery1                           avgt    5     7.036 ±  0.076  ms/op
QueryBenchmark.pathExpressionQuery2                           avgt    5     1.193 ±  0.001  ms/op
QueryBenchmark.query_distinct_predicates                      avgt    5    75.763 ±  0.253  ms/op
QueryBenchmark.simple_filter_not                              avgt    5     3.430 ±  0.005  ms/op
```

## After
```
Benchmark                                    Mode  Cnt   Score   Error  Units
ParallelQueryBenchmark.mixedQueriesAndReads  avgt   15  28.498 ± 0.103  ms/op
```

```
Benchmark                                                     Mode  Cnt     Score    Error  Units
QueryBenchmark.complexQuery                                   avgt    5     1.671 ±  0.025  ms/op
QueryBenchmark.different_datasets_with_similar_distributions  avgt    5     0.807 ±  0.003  ms/op
QueryBenchmark.groupByQuery                                   avgt    5     1.087 ±  0.006  ms/op
QueryBenchmark.long_chain                                     avgt    5   289.228 ±  1.839  ms/op
QueryBenchmark.lots_of_optional                               avgt    5    57.633 ±  0.139  ms/op
QueryBenchmark.minus                                          avgt    5  2447.364 ± 19.308  ms/op
QueryBenchmark.nested_optionals                               avgt    5    81.451 ±  0.158  ms/op
QueryBenchmark.pathExpressionQuery1                           avgt    5     7.005 ±  0.043  ms/op
QueryBenchmark.pathExpressionQuery2                           avgt    5     1.186 ±  0.004  ms/op
QueryBenchmark.query_distinct_predicates                      avgt    5    76.159 ±  0.445  ms/op
QueryBenchmark.simple_filter_not                              avgt    5     2.843 ±  0.004  ms/op
```